+++ client/src/components/QuestionsAnswers/Answer.jsx
@@ -14,6 +14,9 @@ const Answer = (props) => {
         if (res.status !== 200) {
           setHelpful(helpful - 1);
         }
+      })
+      .catch(err => {
+        console.error(err);
       });
   };
 
@@ -33,22 +36,21 @@ const Answer = (props) => {
     <div className="qa-answer-container">
       <span className="qa-body">{answer.body}</span>
       <div className="qa-answer-body">
+        {
+          answer.photos.map((photo) => {
+            return <img
+              className="answer-photo"
+              key={photo.id}
+              src={photo.url ? photo.url : ''}
+            />;
+          })}
+        <div></div>
         <span className="qa-r2  qa-border">By {isSeller()} on {answer.date.slice(0, 10)}</span>
         <span className="qa-r2  "> Helpful? </span>
         <button className="qa-r2 qa-span-btn  qa-border" disabled={helpful > answer.helpfulness} onClick={() => markAsHelpful(answer.answer_id)}> Yes </button>
         <span className="qa-r2 qa-border"> ({helpful}) </span>
-        <span className="qa-r2 qa-question qa-r2 qa-report qa-border" onClick={() => handleReport(item.question_id)}> {isReported}</span>
-        <button className="qa-r2 qa-span-btn qa-add-answer-btn " onClick={() => answerModalHandler(item)}>Add Answer</button>
+        <span className="qa-r2 qa-question qa-r2 qa-report" onClick={() => handleReport(item.question_id)}> {isReported}</span>
       </div>
-      {
-        answer.photos.map((photo) => {
-          return <img
-            className="answer-photo"
-            key={photo.id}
-            src={photo.url ? photo.url : ''}
-          />;
-        })}
-      <div></div>
       <div/>
     </div>
   );
